Refactor gates
