+++ src/qibo/core/gates.py
@@ -29,22 +29,6 @@ class Y(BackendGate, abstract_gates.Y):
         return K.reshape(state, shape)
 
 
-class Align(BackendGate, abstract_gates.Align):
-
-    def __init__(self, *q):
-        BackendGate.__init__(self)
-        abstract_gates.Align.__init__(self, *q)
-
-    def _construct_unitary(self):
-        return K.eye(2 ** len(self.target_qubits))
-
-    def _state_vector_call(self, state):
-        return state
-
-    def _density_matrix_call(self, state):
-        return state
-
-
 class M(BackendGate, abstract_gates.M):
     from qibo.core import measurements, states
 
@@ -159,338 +143,6 @@ class M(BackendGate, abstract_gates.M):
                                         "for more than one shots.")
             return getattr(self, self._active_call)(state)
         return self.result
-
-
-class RX(MatrixGate, abstract_gates.RX):
-
-    def __init__(self, q, theta, trainable=True):
-        MatrixGate.__init__(self)
-        abstract_gates.RX.__init__(self, q, theta, trainable)
-
-    def _construct_unitary(self):
-        theta = self.parameters
-        if isinstance(theta, K.native_types): # pragma: no cover
-            p = K
-            theta = K.cast(theta)
-        else:
-            p = K.qnp
-        cos, isin = p.cos(theta / 2.0) + 0j, -1j * p.sin(theta / 2.0)
-        return K.cast([[cos, isin], [isin, cos]])
-
-
-class RY(MatrixGate, abstract_gates.RY):
-
-    def __init__(self, q, theta, trainable=True):
-        MatrixGate.__init__(self)
-        abstract_gates.RY.__init__(self, q, theta, trainable)
-
-    def _construct_unitary(self):
-        theta = self.parameters
-        if isinstance(theta, K.native_types):
-            p = K
-            theta = K.cast(theta)
-        else:
-            p = K.qnp
-        cos, sin = p.cos(theta / 2.0), p.sin(theta / 2.0)
-        return K.cast([[cos, -sin], [sin, cos]])
-
-
-class RZ(MatrixGate, abstract_gates.RZ):
-
-    def __init__(self, q, theta, trainable=True):
-        MatrixGate.__init__(self)
-        abstract_gates.RZ.__init__(self, q, theta, trainable)
-
-    def _construct_unitary(self):
-        if isinstance(self.parameters, K.native_types): # pragma: no cover
-            p = K
-            theta = K.cast(self.parameters)
-        else:
-            p = K.qnp
-            theta = self.parameters
-        phase = p.exp(1j * theta / 2.0)
-        return K.cast(p.diag([p.conj(phase), phase]))
-
-
-class U1(MatrixGate, abstract_gates.U1):
-
-    def __init__(self, q, theta, trainable=True):
-        MatrixGate.__init__(self)
-        abstract_gates.U1.__init__(self, q, theta, trainable)
-
-    @property
-    def custom_op_matrix(self):
-        if self._custom_op_matrix is None:
-            self._custom_op_matrix = K.qnp.exp(1j * self.parameters)
-        return self._custom_op_matrix
-
-    def _construct_unitary(self):
-        if isinstance(self.parameters, K.native_types): # pragma: no cover
-            p = K
-            theta = K.cast(self.parameters)
-        else:
-            p = K.qnp
-            theta = self.parameters
-        return p.diag([1, p.exp(1j * theta)])
-
-
-class U2(MatrixGate, abstract_gates.U2):
-
-    def __init__(self, q, phi, lam, trainable=True):
-        MatrixGate.__init__(self)
-        abstract_gates.U2.__init__(self, q, phi, lam, trainable)
-
-    def _construct_unitary(self):
-        phi, lam = self.parameters
-        if isinstance(phi, K.native_types) or isinstance(lam, K.native_types): # pragma: no cover
-            p = K
-        else:
-            p = K.qnp
-        eplus = p.exp(1j * (phi + lam) / 2.0)
-        eminus = p.exp(1j * (phi - lam) / 2.0)
-        return K.cast([[p.conj(eplus), - p.conj(eminus)],
-                       [eminus, eplus]]) / p.sqrt(2)
-
-
-class U3(MatrixGate, abstract_gates.U3):
-
-    def __init__(self, q, theta, phi, lam, trainable=True):
-        MatrixGate.__init__(self)
-        abstract_gates.U3.__init__(self, q, theta, phi, lam, trainable)
-
-    def _construct_unitary(self):
-        theta, phi, lam = self.parameters
-        if isinstance(theta, K.native_types) or isinstance(phi, K.native_types) or isinstance(lam, K.native_types): # pragma: no cover
-            p = K
-        else:
-            p = K.qnp
-        cost, sint = p.cos(theta / 2), p.sin(theta / 2)
-        eplus, eminus = p.exp(1j * (phi + lam) / 2.0), p.exp(1j * (phi - lam) / 2.0)
-        return K.cast([[p.conj(eplus) * cost, - p.conj(eminus) * sint],
-                       [eminus * sint, eplus * cost]])
-
-
-class CNOT(BackendGate, abstract_gates.CNOT):
-
-    def __init__(self, q0, q1):
-        BackendGate.__init__(self)
-        abstract_gates.CNOT.__init__(self, q0, q1)
-
-    def _construct_unitary(self):
-        return K.matrices.CNOT
-
-
-class CZ(BackendGate, abstract_gates.CZ):
-
-    def __init__(self, q0, q1):
-        BackendGate.__init__(self)
-        abstract_gates.CZ.__init__(self, q0, q1)
-
-    def _construct_unitary(self):
-        return K.matrices.CZ
-
-
-class _CUn_(MatrixGate):
-    base = U1
-
-    def __init__(self, q0, q1, **params):
-        MatrixGate.__init__(self)
-        cbase = "C{}".format(self.base.__name__)
-        getattr(abstract_gates, cbase).__init__(self, q0, q1, **params)
-
-    def _construct_unitary(self):
-        return MatrixGate._control_unitary(self.base._construct_unitary(self))
-
-    @property
-    def custom_op_matrix(self):
-        if self._custom_op_matrix is None:
-            self._custom_op_matrix = self.base._construct_unitary(self)
-        return self._custom_op_matrix
-
-
-class CRX(_CUn_, abstract_gates.CRX):
-    base = RX
-
-    def __init__(self, q0, q1, theta, trainable=True):
-        _CUn_.__init__(self, q0, q1, theta=theta, trainable=trainable)
-
-
-class CRY(_CUn_, abstract_gates.CRY):
-    base = RY
-
-    def __init__(self, q0, q1, theta, trainable=True):
-        _CUn_.__init__(self, q0, q1, theta=theta, trainable=trainable)
-
-
-class CRZ(_CUn_, abstract_gates.CRZ):
-    base = RZ
-
-    def __init__(self, q0, q1, theta, trainable=True):
-        _CUn_.__init__(self, q0, q1, theta=theta, trainable=trainable)
-
-
-class CU1(_CUn_, abstract_gates.CU1):
-    base = U1
-
-    def __init__(self, q0, q1, theta, trainable=True):
-        _CUn_.__init__(self, q0, q1, theta=theta, trainable=trainable)
-
-    @property
-    def custom_op_matrix(self):
-        if self._custom_op_matrix is None:
-            self._custom_op_matrix = K.qnp.exp(1j * self.parameters)
-        return self._custom_op_matrix
-
-
-class CU2(_CUn_, abstract_gates.CU2):
-    base = U2
-
-    def __init__(self, q0, q1, phi, lam, trainable=True):
-        _CUn_.__init__(self, q0, q1, phi=phi, lam=lam, trainable=trainable)
-
-
-class CU3(_CUn_, abstract_gates.CU3):
-    base = U3
-
-    def __init__(self, q0, q1, theta, phi, lam, trainable=True):
-        _CUn_.__init__(self, q0, q1, theta=theta, phi=phi, lam=lam,
-                       trainable=trainable)
-
-
-class SWAP(BackendGate, abstract_gates.SWAP):
-
-    def __init__(self, q0, q1):
-        BackendGate.__init__(self)
-        abstract_gates.SWAP.__init__(self, q0, q1)
-
-    def _construct_unitary(self):
-        return K.matrices.SWAP
-
-
-class FSWAP(MatrixGate, abstract_gates.FSWAP):
-
-    def __init__(self, q0, q1):
-        BackendGate.__init__(self)
-        abstract_gates.FSWAP.__init__(self, q0, q1)
-
-    def _construct_unitary(self):
-        return K.matrices.FSWAP
-
-
-class fSim(MatrixGate, abstract_gates.fSim):
-
-    def __init__(self, q0, q1, theta, phi, trainable=True):
-        MatrixGate.__init__(self)
-        abstract_gates.fSim.__init__(self, q0, q1, theta, phi, trainable)
-
-    @property
-    def custom_op_matrix(self):
-        if self._custom_op_matrix is None:
-            theta, phi = self.parameters
-            cos, isin = K.qnp.cos(theta) + 0j, -1j * K.qnp.sin(theta)
-            phase = K.qnp.exp(-1j * phi)
-            self._custom_op_matrix = K.cast([cos, isin, isin, cos, phase])
-        return self._custom_op_matrix
-
-    def _construct_unitary(self):
-        theta, phi = self.parameters
-        if isinstance(theta, K.native_types) or isinstance(phi, K.native_types): # pragma: no cover
-            p = K
-        else:
-            p = K.qnp
-        cos, isin = p.cos(theta), -1j * p.sin(theta)
-        matrix = p.eye(4)
-        matrix[1, 1], matrix[2, 2] = cos, cos
-        matrix[1, 2], matrix[2, 1] = isin, isin
-        matrix[3, 3] = p.exp(-1j * phi)
-        return K.cast(matrix)
-
-
-class GeneralizedfSim(MatrixGate, abstract_gates.GeneralizedfSim):
-
-    def __init__(self, q0, q1, unitary, phi, trainable=True):
-        BackendGate.__init__(self)
-        abstract_gates.GeneralizedfSim.__init__(self, q0, q1, unitary, phi, trainable)
-
-    @property
-    def custom_op_matrix(self):
-        if self._custom_op_matrix is None:
-            unitary, phi = self.parameters
-            matrix = K.qnp.zeros(5)
-            matrix[:4] = K.qnp.reshape(unitary, (4,))
-            matrix[4] = K.qnp.exp(-1j * phi)
-            self._custom_op_matrix = K.cast(matrix)
-        return self._custom_op_matrix
-
-    def _construct_unitary(self):
-        unitary, phi = self.parameters
-        if isinstance(unitary, K.native_types) or isinstance(phi, K.native_types): # pragma: no cover
-            p = K
-        else:
-            p = K.qnp
-        matrix = p.eye(4)
-        matrix[1:3, 1:3] = p.reshape(unitary, (2, 2))
-        matrix[3, 3] = p.exp(-1j * phi)
-        return K.cast(matrix)
-
-    def _dagger(self) -> "GeneralizedfSim":
-        unitary, phi = self.parameters
-        if isinstance(unitary, K.native_types):
-            ud = K.conj(K.transpose(unitary))
-        else:
-            ud = unitary.conj().T
-        q0, q1 = self.target_qubits
-        return self.__class__(q0, q1, ud, -phi)
-
-
-class TOFFOLI(BackendGate, abstract_gates.TOFFOLI):
-
-    def __init__(self, q0, q1, q2):
-        BackendGate.__init__(self)
-        abstract_gates.TOFFOLI.__init__(self, q0, q1, q2)
-
-    def _construct_unitary(self):
-        return K.matrices.TOFFOLI
-
-    @property
-    def matrix(self):
-        if self._matrix is None:
-            self._matrix = self._construct_unitary()
-        return self._matrix
-
-
-class Unitary(MatrixGate, abstract_gates.Unitary):
-
-    def __init__(self, unitary, *q, trainable=True, name: Optional[str] = None):
-        if not isinstance(unitary, K.tensor_types):
-            raise_error(TypeError, "Unknown type {} of unitary matrix."
-                                   "".format(type(unitary)))
-        MatrixGate.__init__(self)
-        abstract_gates.Unitary.__init__(self, unitary, *q, trainable=trainable, name=name)
-        n = len(self.target_qubits)
-
-    def _construct_unitary(self):
-        return self.parameters
-
-    def _dagger(self) -> "Unitary":
-        ud = K.conj(K.transpose(self.parameters))
-        return self.__class__(ud, *self.target_qubits, **self.init_kwargs)
-
-    @ParametrizedGate.parameters.setter
-    def parameters(self, x):
-        x = K.cast(x)
-        shape = tuple(x.shape)
-        if len(shape) > 2 and shape[0] == 1:
-            shape = shape[1:]
-            x = K.squeeze(x, axis=0)
-        true_shape = (2 ** self.rank, 2 ** self.rank)
-        if shape == (2 ** (2 * self.rank),):
-            x = K.reshape(x, true_shape)
-        elif shape != true_shape:
-            raise_error(ValueError, "Invalid shape {} of unitary matrix "
-                                    "acting on {} target qubits."
-                                    "".format(shape, self.rank))
-        ParametrizedGate.parameters.fset(self, x) # pylint: disable=no-member
 
 
 class VariationalLayer(BackendGate, abstract_gates.VariationalLayer):
