+++ src/qibo/gates/__init__.py
@@ -1,0 +1,2 @@
+from qibo.gates.gates import *
+from qibo.gates.channels import *
