+++ src/qibo/tests/test_gates_channels.py
